Created tests for diagnosis functionality

+++ src/main/java/com/jupiter/asclepi/core/model/response/disease/DiseaseHistoryInfo.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Value
 public class DiseaseHistoryInfo {
     GetDiseaseHistoryRequest diseaseHistory;
-    List<BigInteger> diagnosisIds;
+    List<Integer> diagnosisIds;
     Integer doctorId;
 
     // todo created by and created when
